MyPkg

Documentation for MyPkg.

introduction

用Simpson积分计算公式计算被积函数f的积分区间为[a, b]定积分

content

两个pages的目录

MyPkg

introduction

content

function

Index

Algorithms

The Rejection Method

Direct Method for systems with delays

Next Reaction Method for systems with delays

练习page

Formula

Cross Reference

URL

function

MyPkg.Simpson — Method

function Simpson(f, a, b)

Calculate the integral f function from x to y by Simpson formula

Here's Simpson expression:

`Simpson(f, a, b) = \frac{(b-a)}{6} * (f(a) + 4*f(\frac{(a+b)}{2}) + f(b))

Arguments

f::Express: integrand function

a::Float: left boundary of integral interval of definite integral

b::Float: right boundary of integral interval of definite integral

Examples

julia> Simpson(2*x, 1, 2)
4

MyPkg.find_num_in_vec — Method

function find_num_in_vec(A::Vector, position_index::Vector{Int64}, x)

Find the number of values which in each vector elements equal to x according to the corresponding index position specified by the element in the position_index vector in the given vetcer A.

Examples

julia> A =  [[0.09,0.09,0.1],[0.3,0.09,0.1],[0.09]]
3-element Vector{Vector{Float64}}:
 [0.09, 0.09, 0.1]
 [0.3, 0.09, 0.1]
 [0.09]

julia> position_index =  [1,2,3]
3-element Vector{Int64}:
 1
 2
 3

julia> find_num_in_vec(A::Vector, position_index::Vector{Int64}, 0.09)
3-element Vector{Int64}:
2
1
1

Index

所有函数的索引

MyPkg.Simpson

MyPkg.find_num_in_vec